test: update unit test for mark status to use resource creation from existing
Retaining the usage of function 'FromExisting' is necessary.

One might argue that this is not a really good approach. Yes, in a certain
cases it is not a good approach. An example in which we no longer use this function
is when we are doing resource deployment, which requires update status.

However, there are certain cases where we still need this function
as a way for shortcut in setting status. An obvious example is
when we are fetching resource in the database. The way we store it
is in a string format, thus we need to convert it. Nonetheless, we don't
know what status it is. And during the resource initialiation, it is initialized
with 'unknown' status. In order to 'reach' the status from database, we have to
perform some actions on marking statuses so it will become the same.
This is tricky, and even the tricky part is we don't know how many steps
we have to perform.

As a side note, we should limit its usage as much as possible.